+++ app/src/androidTest/java/com/example/petmovieproject/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.mypetmovieproject
+package com.example.petmovieproject
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
