+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.mypetmovieproject">
+    package="com.example.petmovieproject">
 
     <application
         android:allowBackup="true"
@@ -8,9 +8,9 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.MyPetMovieProject"
+        android:theme="@style/Theme.PetMovieProject"
         android:fullBackupContent="true">
-        <activity android:name=".MainActivity">
+        <activity android:name="com.example.petmovieproject.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
