--- app/src/main/java/com/example/mypetmovieproject/ClickListener.kt
@@ -1,6 +1,0 @@
-package com.example.mypetmovieproject
-
-interface ClickListener {
-    fun openMovieDetails()
-    fun back()
-}
