+++ app/src/main/java/com/example/petmovieproject/ClickListener.kt
@@ -1,0 +1,6 @@
+package com.example.petmovieproject
+
+internal interface ClickListener {
+    fun openMovieDetails()
+    fun back()
+}
