+++ app/src/test/java/com/example/petmovieproject/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.mypetmovieproject
+package com.example.petmovieproject
 
 import org.junit.Test
